Fix SP1-OO# : move repository package
