--- src/main/java/com/alkemy/wallet/controller/repository/UserRepository.java
@@ -1,5 +1,0 @@
-package com.alkemy.wallet.controller.repository;
-
-public interface UserRepository {
-
-}
