+++ src/main/java/com/alkemy/wallet/repository/TransactionRepository.java
@@ -1,0 +1,5 @@
+package com.alkemy.wallet.repository;
+
+public interface TransactionRepository {
+
+}
